+++ src/basic/App.tsx
@@ -1,5 +1,6 @@
 import { CartItem, ProductWithUI } from '../types';
-import { SearchBar } from './components/ui';
+import { SearchBar } from './components/common';
+import { NotificationList } from './components/notification';
 import { useCart, useNotification, useProduct, useSearch, useToggle } from './hooks';
 import { AdminPage } from './pages/AdminPage';
 import { CartPage } from './pages/CartPage';
@@ -38,7 +39,7 @@ export default function App() {
 
   return (
     <div className="min-h-screen bg-gray-50">
-      <NotificationToast notifications={notifications} onDismiss={addNotification} />
+      <NotificationList notifications={notifications} onDismiss={addNotification} />
 
       <Header
         isAdmin={isAdmin}
@@ -72,53 +73,6 @@ export default function App() {
         completeOrder={completeOrder}
         addNotification={addNotification}
       />
-    </div>
-  );
-}
-
-interface NotificationToastProps {
-  notifications: Array<{
-    id: string;
-    message: string;
-    type: 'error' | 'success' | 'warning';
-  }>;
-  onDismiss: (message: string, type: 'error' | 'success' | 'warning') => void;
-}
-
-function NotificationToast({ notifications, onDismiss }: NotificationToastProps) {
-  if (notifications.length === 0) {
-    return null;
-  }
-
-  return (
-    <div className="fixed top-20 right-4 z-50 space-y-2 max-w-sm">
-      {notifications.map((notif) => (
-        <div
-          key={notif.id}
-          className={`p-4 rounded-md shadow-md text-white flex justify-between items-center ${
-            notif.type === 'error'
-              ? 'bg-red-600'
-              : notif.type === 'warning'
-              ? 'bg-yellow-600'
-              : 'bg-green-600'
-          }`}
-        >
-          <span className="mr-2">{notif.message}</span>
-          <button
-            onClick={() => onDismiss(notif.message, notif.type)}
-            className="text-white hover:text-gray-200"
-          >
-            <svg className="w-4 h-4" fill="none" stroke="currentColor" viewBox="0 0 24 24">
-              <path
-                strokeLinecap="round"
-                strokeLinejoin="round"
-                strokeWidth={2}
-                d="M6 18L18 6M6 6l12 12"
-              />
-            </svg>
-          </button>
-        </div>
-      ))}
     </div>
   );
 }
