+++ src/basic/components/coupon/index.ts
@@ -1,0 +1,1 @@
+export * from './CouponList';
