+++ src/basic/components/index.ts
@@ -1,2 +1,5 @@
-export * from './AdminPage';
-export * from './CartPage';
+export * from './product';
+export * from './cart';
+export * from './coupon';
+export * from './notification';
+export * from './common';
