+++ src/basic/components/product/index.ts
@@ -1,0 +1,5 @@
+export * from './ProductList';
+export * from './ProductPrice';
+export * from './ProductStockStatus';
+export * from './ProductDiscountBadge';
+export * from './ProductDiscountInfo';
