--- src/basic/components/ui/index.ts
@@ -1,5 +1,0 @@
-export * from './EmptyState';
-export * from './ProductList';
-export * from './CartList';
-export * from './NotificationList';
-export * from './SearchBar';
